First pass refactor
